adds transition information
a fsm defines a directed graph where the states are the nodes and the
transitions the edges. it is generally useful to expose this information
to client code.
this additional argument is passed after existing arguments to avoid
breaking existing code. all existing tests are unaffected.
a fresh object is created on each call to avoid a mutable shared
reference